test: CPP-230 Allow CCM commands to be run locally
Added ability to not use/link libssh2 in test framework in the event CCM
commands should only be run locally. Local processes are spawned using
libuv (v0.10 or 1.x). The configuration file for the integration tests
was cleaned up with new options added and old ones removed as well as
the removal of multiple configuration files for performing version one
downgrade tests. The test file is now copied to the build directory and
directory for easy execution and settings modification.

The CCM bridge configuration source was also updated to use the
cpp-driver code style guidelines.